Adds a confirm window when a user make a holiday requests and corrects the vizualisation of holiday requests in manager panel.

+++ holiday/src/main/java/org/kernely/holiday/service/HolidayRequestService.java
@@ -177,7 +177,7 @@ public class HolidayRequestService extends AbstractService {
 	 * @return an holidaydetaildto
 	 */
 	@Transactional
-	public List<HolidayDetailDTO> getHolidayRequestDetails(int idRequest) {
+	public List<HolidayDetailDTO> getHolidayRequestDetails(long idRequest) {
 		Query holidayRequest = em.get().createQuery("SELECT h FROM HolidayRequest h WHERE id=:idRequest");
 		holidayRequest.setParameter("idRequest", idRequest);
 		HolidayRequest holiday = (HolidayRequest) holidayRequest.getSingleResult();
@@ -196,7 +196,7 @@ public class HolidayRequestService extends AbstractService {
 	 * @return an holiday detail DTO
 	 */
 	@Transactional
-	public List<HolidayDetailDTO> getHolidayRequestDetailsByOrder(int idRequest) {
+	public List<HolidayDetailDTO> getHolidayRequestDetailsByOrder(long idRequest) {
 		Query holidayRequest = em.get().createQuery("SELECT h FROM HolidayRequest h WHERE id=:idRequest");
 		holidayRequest.setParameter("idRequest", idRequest);
 		HolidayRequest holiday = (HolidayRequest) holidayRequest.getSingleResult();
